+++ src/Semaforo.java
@@ -1,18 +1,18 @@
 
 
 public class Semaforo {
-    short modo; //Color del semaforo 1 verde, 2 amarillo, 4 rojo, 3 apagado
+    int modo; //Color del semaforo 1 verde, 2 amarillo, 4 rojo, 3 apagado
     ConexionExterna
     
-    public Semaforo(short mod) {
+    public Semaforo(int mod) {
         modo = mod;
     }
 
-    public void setModo(short mod) {
+    public void setModo(int mod) {
         modo = mod;
     }
 
-    public short getModo() {
+    public int getModo() {
         return modo;
     }
 }
